requirements.txt
flask
flask-cors
beautifulsoup4
requests
yt-dlp
google-generativeai
python-dotenv
gunicorn